refactor(exercise): folders
