+++ src/modules/exercise/exerciseDetails.module.css
