+++ src/modules/exercise/services/detectPartOfSpeech.ts
