+++ src/modules/exercise/services/showVariants.ts
